Reply: package refactoring 2
