+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reply/command/repository/ReplyCommandRepository.java
@@ -1,0 +1,7 @@
+package org.thisdote.innerjoinus.articlereply.reply.command.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.thisdote.innerjoinus.articlereply.reply.command.aggregate.ReplyCommandEntity;
+
+public interface ReplyCommandRepository extends JpaRepository<ReplyCommandEntity, Integer> {
+}
