--- innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reply/command/repository/ReplyRepository.java
@@ -1,8 +1,0 @@
-package org.thisdote.innerjoinus.articlereply.reply.command.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.thisdote.communityfordevelopers.reply.command.entity.Reply;
-
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
-
-}
