+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reply/command/vo/ReplyCommandVo.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.articlereply.reply.command.vo;
+
+public class ReplyCommandVo {
+}
