+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reply/query/aggregate/ReplyQueryEntity.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.articlereply.reply.query.aggregate;
+
+public class ReplyQueryEntity {
+}
