+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reply/query/controller/ReplyQueryController.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.articlereply.reply.query.controller;
+
+public class ReplyQueryController {
+}
