+++ innerjoinus-article-reply-source/src/main/resources/application.yml
@@ -3,26 +3,28 @@ server:
 
 spring:
   application:
-    name: swcamp-user-service
+    name: innerjoinus-articlereply
+
   datasource:
     driver-class-name: org.mariadb.jdbc.Driver
-    url: jdbc:mariadb://127.0.0.1:3306/menudb
+    url: jdbc:mariadb://localhost:3306/communityfordeveloper
     username: swcamp
     password: swcamp
 
-jpa:
-  hibernate:
-    ddl-auto: create
-  show-sql: true
-  database: mysql
-  properties:
+  jpa:
     hibernate:
-      '[format_sql]': true
+      ddl-auto: create
+    show-sql: true
+    database: mysql
+    properties:
+      hibernate:
+        '[format_sql]': true
+
 eureka:
   instance:
     instance-id: ${spring.application.name}:${spring.application.instance_id:${random.value}}
   client:
-    fetch-registry: true
     register-with-eureka: true
+    fetch-registry: true
     service-url:
       defaultZone: http://localhost:8888/eureka
