+++ innerjoinus-article-reply-source/src/main/resources/org/thisdote/innerjoinus/articlereply/article/ArticleMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org/thisdote/innerjoinus/articlereply/article/repository/ArticleMapper">
+<mapper namespace="org.thisdote.innerjoinus.articlereply.article.query.repository.ArticleMapper">
 
     <resultMap id="articleResultMap" type="org/thisdote/innerjoinus/articlereply/article/dto/ArticleDTO">
         <id property="articleId" column="USER_CODE"/>
