+++ innerjoinus-article-reply-source/src/main/resources/org/thisdote/innerjoinus/articlereply/reply/ReplyMapper.xml
@@ -3,9 +3,9 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.thisdote.communityfordevelopers.reply.query.ReplyMapper">
+<mapper namespace="org.thisdote.innerjoinus.articlereply.reply.query.repository.ReplyMapper">
 
-    <resultMap id="replyResultMap" type="org.thisdote.communityfordevelopers.reply.dto.ReplyDTO">
+    <resultMap id="replyResultMap" type="org.thisdote.innerjoinus.articlereply.reply.dto.ReplyDTO">
         <id property="replyId" column="reply_id"/>
         <result property="replyReportStatus" column="reply_report_status"/>
         <result property="replyContent" column="reply_content"/>
