+++ innerjoinus-article-reply-source/src/test/java/org/thisdote/articlereply/InnerjoinusArticleReplySourceApplicationTestsEntity.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class InnerjoinusArticleReplySourceApplicationTests {
+class InnerjoinusArticleReplySourceApplicationTestsEntity {
 
     @Test
     void contextLoads() {
